+++ frontend/src/classes/BtrixElement.ts
@@ -1,9 +1,9 @@
 import { TailwindElement } from "./TailwindElement";
 
 import { APIController } from "@/controllers/api";
-import { LocalizeController } from "@/controllers/localize";
 import { NavigateController } from "@/controllers/navigate";
 import { NotifyController } from "@/controllers/notify";
+import localize from "@/utils/localize";
 import appState, { use } from "@/utils/state";
 
 export class BtrixElement extends TailwindElement {
@@ -14,7 +14,7 @@ export class BtrixElement extends TailwindElement {
   readonly api = new APIController(this);
   readonly notify = new NotifyController(this);
   readonly navigate = new NavigateController(this);
-  readonly localize = new LocalizeController(this);
+  readonly localize = localize;
 
   protected get authState() {
     return this.appState.auth;
