+++ frontend/src/features/archived-items/archived-item-list.ts
@@ -13,11 +13,10 @@ import { ifDefined } from "lit/directives/if-defined.js";
 import { CrawlStatus } from "./crawl-status";
 
 import { TailwindElement } from "@/classes/TailwindElement";
-import { LocalizeController } from "@/controllers/localize";
 import { NavigateController } from "@/controllers/navigate";
 import { ReviewStatus, type ArchivedItem } from "@/types/crawler";
 import { renderName } from "@/utils/crawler";
-import { formatISODateString, formatNumber } from "@/utils/localization";
+import { formatISODateString, formatNumber } from "@/utils/localize";
 
 export type CheckboxChangeEventDetail = {
   checked: boolean;
@@ -77,7 +76,6 @@ export class ArchivedItemListItem extends TailwindElement {
   private readonly rowLink?: HTMLAnchorElement;
 
   private readonly navigate = new NavigateController(this);
-  private readonly localize = new LocalizeController(this);
 
   render() {
     if (!this.item) return;
@@ -231,7 +229,6 @@ export class ArchivedItemListItem extends TailwindElement {
             hoist
           >
             <sl-format-date
-              lang=${this.localize.activeLanguage}
               class="truncate"
               date=${this.item.finished}
               month="2-digit"
@@ -251,7 +248,6 @@ export class ArchivedItemListItem extends TailwindElement {
             @click=${this.onTooltipClick}
           >
             <sl-format-bytes
-              lang=${this.localize.activeLanguage}
               class="truncate"
               value=${this.item.fileSize || 0}
               display="narrow"
