+++ frontend/src/features/archived-items/crawl-list.ts
@@ -24,11 +24,10 @@ import { ifDefined } from "lit/directives/if-defined.js";
 import { TailwindElement } from "@/classes/TailwindElement";
 import type { OverflowDropdown } from "@/components/ui/overflow-dropdown";
 import { RelativeDuration } from "@/components/ui/relative-duration";
-import { LocalizeController } from "@/controllers/localize";
 import { NavigateController } from "@/controllers/navigate";
 import type { Crawl } from "@/types/crawler";
 import { renderName } from "@/utils/crawler";
-import { formatNumber } from "@/utils/localization";
+import { formatNumber } from "@/utils/localize";
 import { pluralOf } from "@/utils/pluralize";
 
 const formatNumberCompact = (v: number) =>
@@ -72,7 +71,6 @@ export class CrawlListItem extends TailwindElement {
   dropdownMenu!: OverflowDropdown;
 
   private readonly navigate = new NavigateController(this);
-  private readonly localize = new LocalizeController(this);
 
   render() {
     if (!this.crawl) return;
@@ -84,7 +82,6 @@ export class CrawlListItem extends TailwindElement {
           ${this.safeRender(
             (crawl) => html`
               <sl-format-date
-                lang=${this.localize.activeLanguage}
                 date=${crawl.started}
                 month="2-digit"
                 day="2-digit"
@@ -155,7 +152,6 @@ export class CrawlListItem extends TailwindElement {
                 ${this.safeRender(
                   (crawl) => html`
                     <sl-format-date
-                      lang=${this.localize.activeLanguage}
                       date=${crawl.started}
                       month="2-digit"
                       day="2-digit"
@@ -172,7 +168,6 @@ export class CrawlListItem extends TailwindElement {
             crawl.finished
               ? html`
                   <sl-format-date
-                    lang=${this.localize.activeLanguage}
                     date=${crawl.finished}
                     month="2-digit"
                     day="2-digit"
@@ -198,7 +193,6 @@ export class CrawlListItem extends TailwindElement {
         </btrix-table-cell>
         <btrix-table-cell>
           <sl-format-bytes
-            lang=${this.localize.activeLanguage}
             value=${this.crawl.fileSize || 0}
             display="narrow"
           ></sl-format-bytes>
