+++ frontend/src/features/archived-items/crawl-logs.ts
@@ -2,7 +2,6 @@ import { localized, msg } from "@lit/localize";
 import { css, html, LitElement } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 
-import { LocalizeController } from "@/controllers/localize";
 import type { APIPaginatedList } from "@/types/api";
 import { truncate } from "@/utils/css";
 
@@ -95,8 +94,6 @@ export class CrawlLogs extends LitElement {
       })
     | null = null;
 
-  private readonly localize = new LocalizeController(this);
-
   render() {
     if (!this.logs) return;
     return html`<btrix-numbered-list>
@@ -126,7 +123,6 @@ export class CrawlLogs extends LitElement {
               <div class="row">
                 <div>
                   <sl-format-date
-                    lang=${this.localize.activeLanguage}
                     date=${log.timestamp}
                     month="2-digit"
                     day="2-digit"
