+++ frontend/src/features/crawl-workflows/workflow-list.ts
@@ -22,12 +22,11 @@ import {
 
 import type { OverflowDropdown } from "@/components/ui/overflow-dropdown";
 import { RelativeDuration } from "@/components/ui/relative-duration";
-import { LocalizeController } from "@/controllers/localize";
 import { NavigateController } from "@/controllers/navigate";
 import type { ListWorkflow } from "@/types/crawler";
 import { humanizeSchedule } from "@/utils/cron";
 import { srOnly, truncate } from "@/utils/css";
-import { formatNumber } from "@/utils/localization";
+import { formatNumber } from "@/utils/localize";
 import { pluralOf } from "@/utils/pluralize";
 
 const formatNumberCompact = (v: number) =>
@@ -216,7 +215,6 @@ export class WorkflowListItem extends LitElement {
   dropdownMenu!: OverflowDropdown;
 
   private readonly navigate = new NavigateController(this);
-  private readonly localize = new LocalizeController(this);
 
   render() {
     const notSpecified = html`<span class="notSpecified" role="presentation"
@@ -273,7 +271,6 @@ export class WorkflowListItem extends LitElement {
           ${this.safeRender((workflow) => {
             if (workflow.lastCrawlTime && workflow.lastCrawlStartTime) {
               return html`<sl-format-date
-                  lang=${this.localize.activeLanguage}
                   date="${workflow.lastRun.toString()}"
                   month="2-digit"
                   day="2-digit"
@@ -315,14 +312,12 @@ export class WorkflowListItem extends LitElement {
               workflow.lastCrawlSize
             ) {
               return html`<sl-format-bytes
-                  lang=${this.localize.activeLanguage}
                   value=${+workflow.totalSize}
                   display="narrow"
                 ></sl-format-bytes>
                 <span class="currCrawlSize">
                   +
                   <sl-format-bytes
-                    lang=${this.localize.activeLanguage}
                     value=${workflow.lastCrawlSize}
                     display="narrow"
                   ></sl-format-bytes>
@@ -330,7 +325,6 @@ export class WorkflowListItem extends LitElement {
             }
             if (workflow.totalSize && workflow.lastCrawlSize) {
               return html`<sl-format-bytes
-                lang=${this.localize.activeLanguage}
                 value=${+workflow.totalSize}
                 display="narrow"
               ></sl-format-bytes>`;
@@ -338,7 +332,6 @@ export class WorkflowListItem extends LitElement {
             if (workflow.isCrawlRunning && workflow.lastCrawlSize) {
               return html`<span class="currCrawlSize">
                 <sl-format-bytes
-                  lang=${this.localize.activeLanguage}
                   value=${workflow.lastCrawlSize}
                   display="narrow"
                 ></sl-format-bytes>
@@ -346,7 +339,6 @@ export class WorkflowListItem extends LitElement {
             }
             if (workflow.totalSize) {
               return html`<sl-format-bytes
-                lang=${this.localize.activeLanguage}
                 value=${+workflow.totalSize}
                 display="narrow"
               ></sl-format-bytes>`;
@@ -372,7 +364,6 @@ export class WorkflowListItem extends LitElement {
           ${this.safeRender(
             (workflow) => html`
               <sl-format-date
-                lang=${this.localize.activeLanguage}
                 date="${workflow.modified}"
                 month="2-digit"
                 day="2-digit"
