+++ frontend/src/pages/org/collection-detail.ts
@@ -19,7 +19,7 @@ import type {
 import type { Collection } from "@/types/collection";
 import type { ArchivedItem, Crawl, Upload } from "@/types/crawler";
 import type { CrawlState } from "@/types/crawlState";
-import { formatNumber } from "@/utils/localization";
+import { formatNumber } from "@/utils/localize";
 import { pluralOf } from "@/utils/pluralize";
 
 const ABORT_REASON_THROTTLE = "throttled";
@@ -475,7 +475,6 @@ export class CollectionDetail extends BtrixElement {
           msg("Total Size"),
           (col) =>
             html`<sl-format-bytes
-              lang=${this.localize.activeLanguage}
               value=${col.totalSize || 0}
               display="narrow"
             ></sl-format-bytes>`,
@@ -489,7 +488,6 @@ export class CollectionDetail extends BtrixElement {
           msg("Last Updated"),
           (col) =>
             html`<sl-format-date
-              lang=${this.localize.activeLanguage}
               date=${col.modified}
               month="2-digit"
               day="2-digit"
