+++ frontend/src/pages/org/workflow-detail.ts
@@ -28,7 +28,7 @@ import {
   isActive,
 } from "@/utils/crawler";
 import { humanizeSchedule } from "@/utils/cron";
-import { formatNumber } from "@/utils/localization";
+import { formatNumber } from "@/utils/localize";
 import { isArchivingDisabled } from "@/utils/orgs";
 
 const SECTIONS = ["crawls", "watch", "settings", "logs"] as const;
@@ -777,7 +777,6 @@ export class WorkflowDetail extends BtrixElement {
           msg("Total Size"),
           (workflow) =>
             html` <sl-format-bytes
-              lang=${this.localize.activeLanguage}
               value=${Number(workflow.totalSize)}
               display="narrow"
             ></sl-format-bytes>`,
@@ -973,7 +972,6 @@ export class WorkflowDetail extends BtrixElement {
         ${this.renderDetailItem(msg("Crawl Size"), () =>
           this.workflow
             ? html`<sl-format-bytes
-                lang=${this.localize.activeLanguage}
                 value=${this.workflow.lastCrawlSize || 0}
                 display="narrow"
               ></sl-format-bytes>`
