+++ frontend/src/utils/LiteElement.ts
@@ -5,6 +5,7 @@ import appState, { use } from "./state";
 import { APIController } from "@/controllers/api";
 import { NavigateController } from "@/controllers/navigate";
 import { NotifyController } from "@/controllers/notify";
+import localize from "@/utils/localize";
 
 export { html };
 
@@ -18,6 +19,7 @@ export default class LiteElement extends LitElement {
   private readonly apiController = new APIController(this);
   private readonly notifyController = new NotifyController(this);
   private readonly navigateController = new NavigateController(this);
+  readonly localize = localize;
 
   protected get authState() {
     return this.appState.auth;
