+++ frontend/src/utils/slugify.ts
@@ -1,11 +1,11 @@
 import slugify from "slugify";
 
-import { getActiveLanguage } from "@/controllers/localize";
+import localize from "./localize";
 
 export default function slugifyStrict(value: string) {
   return slugify(value, {
     strict: true,
     lower: true,
-    locale: getActiveLanguage(),
+    locale: localize.activeLanguage,
   });
 }
